proxy_app/cloud_run/requirements.txt
Flask==2.2.2
google-cloud-aiplatform==1.20.0
protobuf==4.21.12